requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
joblib==1.4.2
numpy==1.26.4
protobuf==5.27.1
python-dotenv==1.0.1
scikit-learn==1.5.0
scipy==1.13.1
threadpoolctl==3.5.0
